Query for geometry and attributes both sides of a boundary.
WARNING: This might have a profound change on what data is
available in the tiles.

Previously, the boundaries came from the lines table, where they
were hard to link back to the relation via the "rels" table, and
didn't provide easy left/right distinction without intensive
efforts to reconstruct the whole boundary topology.

Instead, this works around the issue by extracting the admin area
polygons at matching `admin_level`s, ensures that the boundary of
one is correctly oriented and uses the shared portions of boundary
to reconstruct the original. This _should_ work, as long as it is
done before any simplification steps.