Extract POIs from polygons and filter zoom by area.
Previously, the POIs had just been coming from the
`planet_osm_point` table, but some POIs are represented as
polygons. These are included via their centroid, and the same test
function `mz_calculate_poi_level`.

That function has been modified to adjust the zoom based on the
feature's area. Note that this will probably mean we want to take
another look at the zoom levels that some of these get assigned in
the function and adjust based on the distribution of areas for
those feature types.

Finally, the `mz_calculate_poi_level` function gets called almost
a million times for a mid-zoom tile and, even though it's pretty
quick, that's just too slow. The POI `min_zoom` is reified as a
column in both the points and polygons tables and kept up to date
with a trigger (yuck...). That brings the query time for a
mid-zoom tile down from 4s to about 5ms.