Added OSM-based national park polygons at zooms 4-8.
This adds those polygons, removing the NE ones which were in some
of the queries. The new polygons merge in the `boundary` attribute
and update the `mz_calculate_is_landuse` function to treat
national park polygons as a type of landuse.

It also exposes an attribute for the OSM `operator` tag, which can
be used to distinguish US National Parks Service areas from those
run by the US Forestry Service.

Refs #149.