First attempt at some tests.
These are really 'release tests' rather than unit tests, as they
require a lot of data to be loaded in the database first (at the
moment, it requires North America). A better test suite would load
some minimal static 'fixture' data first.

But this is a nice first step, and we can refine it later.