Update national_park tests
* Remove id specifier from test

The feature under test looks like it's now being merged with another,
and is losing the `id` in the tile.

* Remove assertion for national_park at z7

Because we need to update the positive assertion to no longer contain
the id due to merging with other parks, we tried to remove it from the
negative assertion as well. However, it ends up picking a national park
with a different id. Removing it seems better than leaving a test that
isn't very useful.