Re-simplify roads after merging.
We have already simplified the road network and clipped it to the tile by the time we merge (multi) linestrings for the same road together. The problem is that the line merging algorithm will make a single linestring from a multi-linestring, but won't drop the vertices at the merge.

This change introduces a new post-processing filter to run simplification on a single layer and runs it on the road layer post-merge. This seems to bring the vertex count down significantly.