Started adding Python version of kind calculation.
This meant changing some of the YAML to rewrite SQL fragments as YAML. A corresponding change to the SQL output hasn't yet been made, but is possible to keep compatibility for a while.

Currently, this doesn't fully work. However, it shows a potential way that we could move the kind (and general properties) calculation from SQL to Python in preparation for RAWR tiles or similar.