Start adding a method to create SQL output.
Code to produce an output Python file with functions to create the properties dictionary has been implemented. This adds a similar method for SQL output which can be loaded into the database in the same way as `create-sql-functions.py` used to do, but leveraging the new YAML code.

The general idea is to have both pipelines operate in parallel, and to be able to test them both.