Stop on error.
This makes `psql` and the SQL generation code stop the script if they error out. This is important for halting the import process at the point of error, rather than letting it continue until the integration tests fail. By the time it has got to that point, it can be hard to back-track and find the original source of the error.