Hack theme parks filter to deal with vandalism.
This [changeset](http://www.openstreetmap.org/changeset/49488291) appears to be vandalism, and the second instance of it on that particular feature. The user accounts involved seem to be making similarly suspect edits to other features. I have reported it to DWG, and do not want to start an edit war on those features, so have hacked a temporary work-around by allowing `tourism=Theme Park` in addition to the (overwhelmingly conventional) `tourism=theme_park`.