Implement simplification for fixtures.
Update some tests to have smaller fixtures, as this appeared to be causing problems on CircleCI. This is probably a good idea anyway, since these fixtures were several hundred megabytes and wouldn't have been fun to download. This was done though a combination of more aggressive feature clipping and moving relation information outside of the individual features, where it wouldn't be needlessly repeated.