Use label placement point for representative point, if available.
We generate and store a representative point for the whole geometry in the `__label__` and, later, `mz_label_placement` properties. The geometry is clipped at an earlier point in the processing pipeline, so calculating the representative point again results in some duplicate labels. If the original label placement point is available, we can use that instead.