Include all name variants.
Previous behaviour was to de-duplicate name variants with the same value as the main name tag. However, that breaks name fall-back logic when one doesn't want to fall back to the main name. For example, something tagged with `name=Foo, name:A=Foo, name:B=Bar` would remove the `name:A` tag, meaning that if the name preference display was `A` then `B` then local, it would display `Bar` instead of `Foo`.