Add an extra test case to cover a corner case.
Where a road has `network="X"` and `ref="1;Y2"`, the intention is clearly that the road should be labelled `"X1"` and `"Y2"`. However, this wasn't getting picked up in the absence of _both_ X and Y relations. The new test covers that, and the logic can now "use up" the element's `network` tag to fill in for a missing prefix on the first `ref`.