Chinese road shields.
Added a `shield_text` function to `CountryNetworkLogic`, which allows us to customise how the ref is turned into shield text on per-country basis. Previously, this was handled in `_road_shield_text`, which is still the default, although special cases will get factored out of that and into per-country cases as we go along.