Refactor test matching code to be recursive on lists.
This means we're able to write tests which have lists which contain unicode strings, which can be more readable. These are now converted to UTF-8 encoded `str` during the test, whereas before they were compared `==` which failed.

This meant changing a few tests where we had `None` in lists, but is now more consistent with the rest of the tests; `None` means any, and `type(None)` means actually `None`.