Add first version of a test generator.
The idea is to replace a lot of the external data which we currently have either loaded from OSM or in shapefile fixtures with data that's explicitly in the test file. To make this easier, the `scripts/mktest.py` script can handle loading and transforming data, writing out Python code which can be appended to a test.

For example, to generate a test of the road transformations, one might run `python scripts/mktest.py road --way-id 389741906 --is-in ES --expect '{"shield_text": "A-132", "network": "ES:autonoma"}' >> integration-test/1521-spanish-road-shields.py`. The template is set up in such a way that appending to an existing test results in relatively well-formatted code.

These tests still usually need a bit of cleanup to make sure it's valid Python syntax (e.g: names of test functions), fit with PEP8, and remove unnecessary tags or properties which make the test less readable.