Add cross-junction merging.
This merges lines across junctions (i.e: multiple lines meeting at a point) where the lines continue on a similar bearing. This means that lines can be more aggressively simplified, and result in fewer shape points and smaller file sizes.

Also, drop small line segments part of larger MultiLineStrings, which is similar to how we drop small inners of MultiPolygons. If a single line segment is much smaller than a pixel, it's not really adding anything to the tile, but might still need 2 points to describe it.

Since many geometry operations will make a MultiLineString "simple", and this re-introduces intersections, we also separate features so that MultiLineStrings do not contain individual LineStrings which intersect.