Drop many "detail" tags on minor roads at mid zooms.
On minor roads (only - this doesn't apply to major roads or highways), drop the following from zoom 15 down:

* `colour`
* `motor_vehicle`
* `operator`
* `route`
* `route_name`
* `state`
* `surface`
* `symbol`
* `type`

And the following from zoom 13 down:

* `cutting`
* `embankment`

This helps a lot with road merging, as the fewer distinct combinations of road properties there are, the more roads can be merged into one another.

Also, drop most values of the `access` property from zoom 13 down, keeping `access=no`, `access=private` and `access=official`. The `no` and `private` values are important for knowing which streets are inaccessible, but many other distinctions are unimportant at this zoom level, where streets might be a few pixels long, e.g: `access=customers` (if styled at all) can be made prominent as the view is zoomed in a bit more.

Further, we merge some `access` values together; `emergency`, `military`, `restricted` and `forestry` all map to `access=official`, which indicates that there is some kind of restriction, but allows for geometry merging and value compression through re-use.

These measures together produced a 14% reduction in the number of `minor_road` features on `12/1206/1540`, reflected in a similar drop in the size of the metadata and properties for the MVT tile. Overall, the `roads` layer was 4% smaller (about 3kB for this tile).