Reduce the number of bogus service areas at mid zooms.
The tag `highway=services` appears to get mis-applied to many motoring services stores. Since highway services, which group fuel stations, toilets, restaurants, etc..., are important and highly visible in driving-oriented styles, these mis-tagged points can be conspicuously visible.

This PR applies two filters to boost some genuine `highway=services` to low zooms while trying to keep the mis-tagged items at higher zooms. The first is an area filter, based on observed areas of genuine-looking service areas. The second is based on the name, which often includes a strong hint.

A third signal which is more complex and so wasn't used this time, but might be interesting to add to the mix if this continues to be a problem, is to look at the proximity of the nearest major road and related amenities (parking, fuel, toilets, etc...).