Add script to output all the kinds in the YAML files.
It will also read the `sort_rank` spreadsheet CSVs and output the sort rank information where available.

The output is a table of the layer, kind, start zoom and sort rank (or `None` where those last two aren't available). The "start zoom" is the minimum `min_zoom` at which the feature could possibly appear. For some features where we use an area-based scaling rule, this could theoretically be zero if the feature were large enough - but this doesn't happen in practice. However, it indicates that we could need to put a `clamp` on the range to prevent them from showing up at unexpectedly low zooms. The start zoom can be `None` if the information is obtained externally, e.g: from an external table such as WOF or NE.

Additionally, did some cleanup to the YAML files:

1. The WOF layers didn't explicitly mention their `placetype`, so that has been whitelisted now. In practice, these were the only `placetype`s values in the database, but it's helpful to be able to know that by looking at the YAML and not having to dig into the database.
2. There were many instances of a `tags->` prefix on column names, which is a holdover from when we used to have OSM tables which were partially column-based and partly hstore-based. These aren't needed any more, so I removed them.