Clip to Mercator world bounds before projecting shapefiles.
Previously, we had set an option to make `ogr2ogr` drop points which could not be projected into Mercator coordinates. This generally worked fine, and we ended up with output that looked correct.

However, the 1:110M oceans polygon was being truncated because there wasn't a point at exactly the corner of the Mercator world, and dropping the out-of-range points left a bit of a gap.

We can fix this by clipping (i.e: intersecting) with the Mercator world bounds before projecting. This should ensure that points are within the destination projection without the need to drop them. It's effectively inserting a point at the corner of the Mercator world.