Add incident_id to Traffic Incidents Layer
One of our engineers has identified the following use case leading to this proposal:

In a case where lets say 3 road segments are of different road_kind, The tile should have 3 different features. Using the incident ID as a feature ID breaks the requirement that all features inside a tile must have unique ID. Thus, we decided that in this particular case that is quite rare but still possible, we will assign unique feature id for each feature and add a common incident_id field for all of them.