Check that building height is in a sensible range.
Sometimes, data that doesn't make sense gets into OSM. For example, version 4 of [this building](https://www.openstreetmap.org/way/290535940) has `height=5e+70`, which is many orders of magnitude larger than the observable universe. Clearly this is a typo and causes problems for us when we try to write it out in tiles.

This patch drops values outside a sensible range, which is approximately that of the tallest current structure, the Burj Khalifa.