Assign aerodrome kind_detail to runways.
Aerodrome polygons are (usually) used to map the footprint of an airport, and often have an `aerodrome:type` tag to indicate whether it's an international or domestic airport. We added that as `kind_detail` in #1277, but it's useful to know that information on runways also. This allows us to hide runways of smaller airports and emphasise those of bigger (international) ones.

This extends the intercut / overlap methods to add "where" clauses, so that we can select those features within the layer which are being cut or used for cutting. We can then overlap the aerodrome polygons with the runways (lines in `roads` layer, polygons in `landuse`) to assign the `kind_detail`.

Worth noting; we already had a `kind_detail` on the runway lines in the `roads` layer, so I've called this new attribute `aerodrome_kind_detail` instead.