Add Wikidata queries to static assets
Wikidata can be a very useful source of extra information to extend the data coming from OpenStreetMap. In this case, we're extracting the number of passengers through an airport from Wikidata and will use that to categorize airports into major and minor.

The data is added as a new table, `wikidata`, indexed by the ID (i.e: `'Q[0-9]*'`) and providing `tags` which will be merged into the tags coming from OSM. Since we might do multiple queries, and these might return different tags on overlapping sets of Wikidata IDs, we merge all the Wikidata query results together. This means we should be careful not to use the same attribute names for more than one query.