Override or backfill aerodrome kind_detail
We would like to be able to distinguish between large, important airports and smaller, less important airports in the map. However, the tagging of airports (`kind: aerodrome`) from OSM is somewhat inconsistent. We were previously using the `aerodrome:type`, but this is often missing.

This patch adds several backfilling mechanisms:

1. Using the `aerodrome` tag in addition to `aerodrome:type`. The two tags share many values in practice, so we can use `aerodrome=international` to override `aerodrome:type=*`.
2. Using the `international_flights=yes` tag to override `aerodrome` and `aerodrome:type`.
3. Joining to Wikidata to get the number of passengers the airport serves each year, allowing us to backfill not only very large airports with `kind_detail: international`, but also medium-sized airports where we weren't able to figure out the type from the `aerodrome` and `aerodrome:type` tags (if there were any).

Caveats: We look for airports with more than a million passengers per year to identify "international" airports. It turns out that there are several _domestic_ airports which serve more than a million people a year. For those airports, "international" is probably best interpreted as a description of the size of the airport, rather than the routes flown to/from it.