Include geometry for disputed boundaries and claims.
Until now, we had only been including boundary features from the `planet_osm_polygon` table and nothing at all from the `planet_osm_line` table. However, the disputes and claims are linear features, so we needed to include those.

However, because `osm2pgsql` creates both linear and polygonal features for boundaries (ways and relations), then we must exclude all the linear features that are in `planet_osm_line` _except_ for the disputes and claims. This is further complicated because we change the geometry type during the query, dropping the polygon for its (oriented) boundary.

The way I've done this is to include a new flag, `mz_boundary_from_polygon`, which basically means "this was a polygon, but we transformed it", and this allows us to write YAML that works. However, it's kind of a horrible hack and I mildly regret doing it. A better alternative might have been to leave the polygons in the query (appropriately clipped to an expanded bounding box) and do the boundary transform at a later stage in the stack.

Lots of changes to the integration tests to adjust for the new hacky flag...