Add more viewpoint data from Natural Earth
Adds viewpoint data from the v5.0-pre4 version. This extends disputed information to admin-1 (state) boundaries and populated places. In the case of the boundary lines, the data is the same (`kind:xx` overrides) as for previous 1:10m disputed boundary work.

In the case of populated places, the dispute data is used to set overrides for the `country_capital` and `region_capital` properties, i.e: `country_capital:xx=True/False`.