Backfill region capital status from Natural Earth
To figure out if a place (`city` or `town`) is a region or state capital, we have previously only looked at the OSM `state_capital` tag. However, some region capitals aren't tagged as such in OSM and this leads to changes across the boundary between Natural Earth and OSM data.

This patch back-fills the `region_capital` status from Natural Earth, so that if _either_ the OSM `state_capital` or the Natural Earth `featurecla` have appropriate values, then the `region_capital` flag will be set.