Add dilate / buffer into water and clip against water layer. This means that landcover should match up (modulo simplification) with the water layer and not have some landcover overlap into the water - or massive gaps between the landcover and water.
Add modal pixel filter, which replaces the central pixel of a kernel with the most common pixel over that kernel. This consolidates blocks of pixels and reduces noise in the data, as the landcover raster can look a little like it has been dithered, with a lot of single pixels.

Add a `GDALSieveFilter` pass to remove small blobs. This is necessary because the modal filter can still result in single, stand-alone, pixels and these don't look good when polygonized and put into the tiles. However, the sieve filter alone (without the modal filter) also doesn't look good, as it doesn't consolidate blocks of pixels, leading to odd artefacts where there's a lot of "dithering" (i.e: are these light green pixels on a dark green background, or dark green pixels on a light green background?).